+++ library/src/main/java/bg/devlabs/fullscreenvideoview/playbackspeed/OnPlaybackSpeedSelectedListener.java
@@ -1,0 +1,10 @@
+package bg.devlabs.fullscreenvideoview.playbackspeed;
+
+/**
+ * Created by Slavi Petrov on 18.06.2019
+ * Dev Labs
+ * slavi@devlabs.bg
+ */
+public interface OnPlaybackSpeedSelectedListener {
+    void onSpeedSelected(float speed, String text);
+}
