+++ library/src/main/java/bg/devlabs/fullscreenvideoview/playbackspeed/PlaybackSpeedPopupMenu.java
@@ -1,4 +1,4 @@
-package bg.devlabs.fullscreenvideoview;
+package bg.devlabs.fullscreenvideoview.playbackspeed;
 
 import android.content.Context;
 import android.view.Menu;
@@ -14,16 +14,16 @@ import java.util.Locale;
  * Dev Labs
  * slavi@devlabs.bg
  */
-class PlaybackSpeedPopupMenu extends android.support.v7.widget.PopupMenu {
+public class PlaybackSpeedPopupMenu extends android.support.v7.widget.PopupMenu {
 
     private ArrayList<Float> values = new ArrayList<>(Arrays.asList(0.25f, 0.5f, 0.75f, 1f, 1.25f, 1.5f, 2f));
 
-    PlaybackSpeedPopupMenu(Context context, View anchor) {
+    public PlaybackSpeedPopupMenu(Context context, View anchor) {
         super(context, anchor);
         addMenuButtons();
     }
 
-    public void setOnSpeedSelectedListener(final OnSpeedSelectedListener listener) {
+    public void setOnSpeedSelectedListener(final OnPlaybackSpeedSelectedListener listener) {
         setOnMenuItemClickListener(new OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
@@ -35,7 +35,7 @@ class PlaybackSpeedPopupMenu extends android.support.v7.widget.PopupMenu {
         });
     }
 
-    void setPlaybackSpeedOptions(PlaybackSpeedOptions playbackSpeedOptions) {
+    public void setPlaybackSpeedOptions(PlaybackSpeedOptions playbackSpeedOptions) {
         values.clear();
         values = playbackSpeedOptions.getSpeeds();
         removeMenuButtons();
@@ -54,9 +54,5 @@ class PlaybackSpeedPopupMenu extends android.support.v7.widget.PopupMenu {
             String title = String.format(Locale.getDefault(), "%.2f", values.get(i)) + "x";
             getMenu().add(0, id, Menu.NONE, title);
         }
-    }
-
-    interface OnSpeedSelectedListener {
-        void onSpeedSelected(float speed, String text);
     }
 }
